ci: build Python 3.8 with MacOSX 10.15 SDK from older Xcode
See the inline comment and https://github.com/indygreg/PyOxidizer/issues/373
for more context. The goal of this change is to get rid of the reference to
__darwin_check_fd_set_overflow to restore working Python distributions
for PyOxidizer. This is far from a long-term solution as we need to
handle weak symbols more robustly for Python 3.9+. But it seems
appropriate for Python 3.8 builds, as the introduction of this symbol
was a "regression" when we transitioned to GitHub Actions and apparently
introduced the buggy SDK.